backend/requirements.txt
fastapi
uvicorn
pdfplumber
PyYAML
python-multipart
